init_document 함수를 수정함 (collect_data.sh  재 실행시 mem_statistics.csv 파일 초기화 문제를 수정함)

+++ collect_data.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-#HyunGwan Seo
+#Hyeongwan Seo
 
 #./collect_data.sh stop -> 데이터 수집을 중지함.
 
@@ -44,13 +44,17 @@ function print_data()
 	do
 		get_data
 		echo "$DAY-$MONTH-$YEAR $HOUR:$MINUTE:$SECOND +0009,$WEEK,$YEAR,$MONTH,$DAY,$HOUR,$MINUTE,$SECOND,$MemFree_mb,$Active_mb,$Cached_mb" >> $MEM_STATISTICS
-		sleep 5m	# 데이터 수집 주기 설정 ex) 30s -> 30초, 10m -> 10분, 1h -> 1시간
+		sleep 2s	# 데이터 수집 주기 설정 ex) 30s -> 30초, 10m -> 10분, 1h -> 1시간
 	done
 }
 
 function init_document()
 {
-	echo "Timestamp,Week,Year,Month,Day,Hour,Minute,Second,MemFree,Active,Cached" > $MEM_STATISTICS
+	if [ -e $MEM_STATISTICS ]; then
+		:		#NOP
+	else
+		echo "Timestamp,Week,Year,Month,Day,Hour,Minute,Second,MemFree,Active,Cached" > $MEM_STATISTICS
+	fi
 }
 
 # 1개의 프로세스만 메모리 사용량을 수집하도록 함.
@@ -71,7 +75,7 @@ function process_check()
 
 R_check
 process_check
-init_document  # R에서 추가 t여부 검토
+init_document  # csv 파일이 존재하면 init_document 함수를 실행하지 않도록 해야함
 
 #echo "collect_data.sh 프로세스 개수: $PROCESS_NUM"
 #echo "collect_data.sh 프로세스 번호: `pgrep collect_data.sh`"
